+++ src/Application/Communication/MessageTestResult.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EventFlow\Application\Communication;
+
+final readonly class MessageTestResult
+{
+    public function __construct(public MessageRecord $message, public int $jobId) {}
+}